Minor fixes.
[FIXED]     Missing import statements; don't assume we're building with a certain PCH.
[FIXED]     Correct typing; esp. for 64-bit iOS.

+++ DCIntrospect/DCFrameView.m
@@ -121,7 +121,7 @@
 
 	UIFont *font = [UIFont systemFontOfSize:10.0f];
 
-	float dash[2] = {3, 3};
+	CGFloat dash[2] = {3, 3};
 	CGContextSetLineDash(context, 0, dash, 2);
 
 	// edge->left side
